ready login via google logic, alarm manager for resetting time limitation and exercise status, Confirmation email, accessibilityService rework, Parent screen, devision between parent and child on google auth

+++ app/src/main/java/com/movetoplay/presentation/ui/child_home/Home.kt
@@ -34,9 +34,9 @@ fun Home(
     ) {
         Spacer(modifier = Modifier.height(46.dp))
         TimeUse(
-            availableForDayMinutes = AccessibilityPrefs.remainingTime.toInt(),
-            remainderMinutes = AccessibilityPrefs.dailyLimit.toInt(),
-            addTime = {},
+            availableForDayMinutes = AccessibilityPrefs.remainingTime,
+            remainderMinutes = AccessibilityPrefs.dailyLimit,
+            addTime = { },
             sizeButton = sizeButtonAndIndicators
         )
         Spacer(modifier = Modifier.height(20.dp))
